use plugins to find docs files
