+++ pyproject.toml
@@ -90,6 +90,16 @@ scsetup = "siliconcompiler.utils.showtools:setup"
 [project.entry-points."siliconcompiler.target"]
 targets = "siliconcompiler.targets:get_targets"
 
+[project.entry-points."siliconcompiler.docs"]
+linkcode = "siliconcompiler.sphinx_ext:get_codeurl"
+targets = "siliconcompiler.sphinx_ext:targets"
+flows = "siliconcompiler.sphinx_ext:flows"
+libs = "siliconcompiler.sphinx_ext:libraries"
+pdks = "siliconcompiler.sphinx_ext:pdks"
+tools = "siliconcompiler.sphinx_ext:tools"
+apps = "siliconcompiler.sphinx_ext:apps"
+checklists = "siliconcompiler.sphinx_ext:checklists"
+
 [project.optional-dependencies]
 test = [
     "pytest == 8.3.4",
