+++ siliconcompiler/flows/showflow.py
@@ -1,12 +1,12 @@
 import siliconcompiler
 from siliconcompiler import SiliconCompilerError
-from siliconcompiler.targets import freepdk45_demo
 
 
 ############################################################################
 # DOCS
 ############################################################################
 def make_docs(chip):
+    from siliconcompiler.targets import freepdk45_demo
     chip.use(freepdk45_demo)
     return setup(filetype='gds', showtools=chip._showtools, np=3)
